+++ R/ts_plot.R
@@ -273,7 +273,38 @@ parse_unit <- function(by) {
 }
 
 
-round_time <- function(x, interval = 60, center = TRUE) {
+#' A generic function for rounding date and time values
+#'
+#' @param x A vector of class POSIX or Date.
+#' @param n Unit to round to. Defaults to mins. Numeric values treated
+#'   as seconds. Otherwise this should be one of "mins", "hours", "days",
+#'   "weeks", "months", "years" (plural optional).
+#' @return If POSIXct then POSIX. If date then Date.
+#' @export
+#' @importFrom hms hms
+#' @examples
+#'
+#' ## class posixct
+#' round_time(Sys.time(), "12 hours")
+#'
+#' ## class date
+#' unique(round_time(seq(Sys.Date(), Sys.Date() + 100, "1 day"), "weeks"))
+round_time <- function(x, n) UseMethod("round_time")
+
+#' @export
+round_time.POSIXt <- function(x, n = "mins") {
+  n <- parse_to_secs(n)
+  as.POSIXct(hms::hms(as.numeric(x) %/% n * n))
+}
+
+#' @export
+round_time.Date <- function(x, n = "months") {
+  x <- as.POSIXct(x)
+  as.Date(round_time(x, n))
+}
+
+
+round_time2 <- function(x, interval = 60, center = TRUE) {
   stopifnot(inherits(x, "POSIXct"))
   ## parse interval
   interval <- parse_unit(interval)
@@ -314,4 +345,33 @@ trim_ots <- function(x, f = 1L, l = 1L) {
     return(x)
   }
   x[-c(f, l), ]
+}
+
+
+parse_to_secs <- function(x) {
+  if (is.numeric(x)) {
+    n <- x
+  } else if (grepl("year", x)) {
+    n <- 60 * 60 * 24 * 365
+  } else if (grepl("month", x)) {
+    n <- 60 * 60 * 24 * 30
+  } else if (grepl("week", x)) {
+    n <- 60 * 60 * 24 * 7
+  } else if (grepl("day", x)) {
+    n <- 60 * 60 * 24
+  } else if (grepl("hour", x)) {
+    n <- 60 * 60
+  } else if (grepl("min", x)) {
+    n <- 60
+  } else if (grepl("sec", x)) {
+    n <- 1
+  } else {
+    stop("must express time interval in secs, mins, hours, days, weeks, months, or years",
+      call. = FALSE)
+  }
+  x <- as.double(gsub("[^[:digit:]|\\.]", "", x))
+  if (any(is.na(x), identical(x, ""))) {
+    x <- 1
+  }
+  n * x
 }